Fix promote prega problem in build script in 2.6 branch also
git-svn-id: https://svn.code.sf.net/p/xcat/code/xcat-core/branches/2.6@10486 8638fb3e-16cb-4fca-ae20-7b5d299a9bcd

+++ buildcore.sh
@@ -373,8 +373,10 @@ fi
 if [ ! -e core-snap ]; then
 	ln -s xcat-core core-snap
 fi
-while ! rsync -urLv --delete $CORE $UPLOADUSER,xcat@web.sourceforge.net:htdocs/$YUM/$REL/
-do : ; done
+if [ "$REL" = "devel" -o "$PREGA" != 1 ]; then
+	while ! rsync -urLv --delete $CORE $UPLOADUSER,xcat@web.sourceforge.net:htdocs/$YUM/$REL/
+	do : ; done
+fi
 
 # Upload the individual source RPMs to sourceforge
 while ! rsync -urLv --delete $SRCD $UPLOADUSER,xcat@web.sourceforge.net:htdocs/$YUM/$REL/
